+++ lib/@mp_idx_manager/mp_idx_manager.m
@@ -22,7 +22,7 @@ classdef mp_idx_manager < handle
 %           struct defining the various set types, where the key is the
 %           set type name, which must also be declared as a property in
 %           the object's class, and the value is a string name used for
-%           display purposes.
+%           display purposes or a mp.set_manager object.
 %
 %           E.g.
 %               function obj = def_set_types(obj)
@@ -146,13 +146,21 @@ classdef mp_idx_manager < handle
                         warning(s1.state, 'Octave:classdef-to-struct');
                     end
                     for k = 1:length(props)
-                        obj.(props{k}) = s.(props{k});
+                        if isa(s.(props{k}), 'mp.set_manager')
+                            obj.(props{k}) = s.(props{k}).copy();
+                        else
+                            obj.(props{k}) = s.(props{k});
+                        end
                     end
                 elseif isstruct(s)
                     props = fieldnames(obj);
                     for k = 1:length(props)
                         if isfield(s, props{k})
-                            obj.(props{k}) = s.(props{k});
+                            if isa(s.(props{k}), 'mp.set_manager')
+                                obj.(props{k}) = s.(props{k}).copy();
+                            else
+                                obj.(props{k}) = s.(props{k});
+                            end
                         end
                     end
                 else
@@ -180,23 +188,16 @@ classdef mp_idx_manager < handle
         function obj = init_set_types(obj)
             % Initialize indexing structures for each set type.
 
-            %% base data struct for each type
-            es = struct();
-            ds = struct( ...
-                'idx', struct( ...
-                    'i1', es, ...
-                    'iN', es, ...
-                    'N', es ), ...
-                'N', 0, ...
-                'NS', 0, ...
-                'order', struct( ...
-                    'name', [], ...
-                    'idx', [] ), ...
-                'data', es );
+            %% Can allow def_set_types() to return a struct whose values
+            %% are either char arrays or mp.set_manager objects.
 
             %% initialize each (set_type) field with base data structure
             for f = fieldnames(obj.set_types)'
-                obj.(f{1}) = ds;
+                nis = obj.set_types.(f{1});
+                if ischar(nis)
+                    nis = mp.set_manager(nis);
+                end
+                obj.(f{1}) = nis;
             end
         end
 
@@ -214,41 +215,18 @@ classdef mp_idx_manager < handle
                 warning(s1.state, 'Octave:classdef-to-struct');
             end
             for k = 1:length(props)
-                new_obj.(props{k}) = obj.(props{k});
+                if isa(obj.(props{k}), 'mp.set_manager')
+                    new_obj.(props{k}) = obj.(props{k}).copy();
+                else
+                    new_obj.(props{k}) = obj.(props{k});
+                end
             end
         end
 
         function display_set(obj, stype, sname)
             % Display indexing information for a given set type.
 
-            if nargin < 3
-                sname = obj.set_types.(stype);
-            end
-            st = obj.(stype);    %% data for set type of interest
-            if st.NS
-                fmt = '%-26s %6s %8s %8s %8s\n';
-                fprintf(fmt, sname, 'name', 'i1', 'iN', 'N');
-                fprintf(fmt, repmat('=', 1, length(sname)), '------', '-----', '-----', '------');
-                idx = st.idx;
-                fmt = '%10d:%22s %8d %8d %8d\n';
-                for k = 1:st.NS
-                    name = st.order(k).name;
-                    if isempty(st.order(k).idx)
-                        fprintf(fmt, k, name, idx.i1.(name), idx.iN.(name), idx.N.(name));
-                    else
-                        vsidx = st.order(k).idx;
-                        str = '%d'; for m = 2:length(vsidx), str = [str ',%d']; end
-                        s = substruct('.', name, '()', vsidx);
-                        nname = sprintf(['%s(' str, ')'], name, vsidx{:});
-                        fprintf(fmt, k, nname, ...
-                                subsref(idx.i1, s), subsref(idx.iN, s), subsref(idx.N, s));
-                    end
-                end
-                fmt = sprintf('%%10d = %%s.NS%%%dd = %%s.N\\n\\n', 35-length(stype));
-                fprintf(fmt, st.NS, stype, st.N, stype);
-            else
-                fprintf('%-26s  :  <none>\n', sname);
-            end
+            obj.(stype).display();
         end
 
         obj = add_named_set(obj, set_type, name, idx, N, varargin)